+++ assets/js/dashboard/stats/locations/map.js
@@ -11,7 +11,7 @@ import * as api from '../../api'
 import { navigateToQuery } from '../../query'
 import { withThemeConsumer } from '../../theme-consumer-hoc';
 
-class CountriesMap extends React.Component {
+class Countries extends React.Component {
   constructor(props) {
     super(props)
     this.resizeMap = this.resizeMap.bind(this)
@@ -21,18 +21,9 @@ class CountriesMap extends React.Component {
     this.onVisible = this.onVisible.bind(this)
   }
 
-  onVisible() {
-    this.fetchCountries().then(this.drawMap.bind(this))
-    window.addEventListener('resize', this.resizeMap);
-    if (this.props.timer) this.props.timer.onTick(this.updateCountries.bind(this))
-  }
-
-  componentWillUnmount() {
-    window.removeEventListener('resize', this.resizeMap);
-  }
-
   componentDidUpdate(prevProps) {
     if (this.props.query !== prevProps.query) {
+      // eslint-disable-next-line react/no-did-update-set-state
       this.setState({loading: true, countries: null})
       this.fetchCountries().then(this.drawMap)
     }
@@ -46,18 +37,29 @@ class CountriesMap extends React.Component {
     }
   }
 
+  componentWillUnmount() {
+    window.removeEventListener('resize', this.resizeMap);
+  }
+
+  onVisible() {
+    this.fetchCountries().then(this.drawMap.bind(this))
+    window.addEventListener('resize', this.resizeMap);
+    if (this.props.timer) this.props.timer.onTick(this.updateCountries.bind(this))
+  }
+
   getDataset() {
-    var dataset = {};
+    const dataset = {};
 
     var onlyValues = this.state.countries.map(function(obj){ return obj.visitors });
     var maxValue = Math.max.apply(null, onlyValues);
 
-    var paletteScale = d3.scale.linear()
+    // eslint-disable-next-line no-undef
+    const paletteScale = d3.scale.linear()
       .domain([0,maxValue])
       .range([
         this.props.darkTheme ? "#2e3954" : "#f3ebff",
         this.props.darkTheme ? "#6366f1" : "#a779e9"
-      ]);
+      ])
 
     this.state.countries.forEach(function(item){
       dataset[item.name] = {numberOfThings: item.visitors, fillColor: paletteScale(item.visitors)};
@@ -82,7 +84,7 @@ class CountriesMap extends React.Component {
   }
 
   drawMap() {
-    var dataset = this.getDataset();
+    const dataset = this.getDataset();
     const label = this.props.query.period === 'realtime' ? 'Current visitors' : 'Visitors'
     const defaultFill = this.props.darkTheme ? '#2d3747' : '#f8fafc'
     const highlightFill = this.props.darkTheme ? '#374151' : '#F5F5F5'
@@ -98,16 +100,14 @@ class CountriesMap extends React.Component {
       geographyConfig: {
         borderColor,
         highlightBorderWidth: 2,
-        highlightFillColor: function(geo) {
-          return geo['fillColor'] || highlightFill;
-        },
+        highlightFillColor: (geo) => geo.fillColor || highlightFill,
         highlightBorderColor,
-        popupTemplate: function(geo, data) {
-          if (!data) { return ; }
+        popupTemplate: (geo, data) => {
+          if (!data) { return null; }
           const pluralizedLabel = data.numberOfThings === 1 ? label.slice(0, -1) : label
           return ['<div class="hoverinfo dark:bg-gray-800 dark:shadow-gray-850 dark:border-gray-850 dark:text-gray-200">',
             '<strong>', geo.properties.name, '</strong>',
-            '<br><strong class="dark:text-indigo-400">', numberFormatter(data.numberOfThings), '</strong> ' + pluralizedLabel,
+            '<br><strong class="dark:text-indigo-400">', numberFormatter(data.numberOfThings), '</strong>', pluralizedLabel,
             '</div>'].join('');
         }
       },
@@ -131,18 +131,22 @@ class CountriesMap extends React.Component {
         <span className="text-xs text-gray-500 absolute bottom-4 right-3">IP Geolocation by <a target="_blank" href="https://db-ip.com" rel="noreferrer" className="text-indigo-600">DB-IP</a></span>
       )
     }
+
+    return null
   }
 
   renderBody() {
     if (this.state.countries) {
       return (
-        <React.Fragment>
+        <>
           <div className="mx-auto mt-4" style={{width: '100%', maxWidth: '475px', height: '335px'}} id="map-container"></div>
           <MoreLink site={this.props.site} list={this.state.countries} endpoint="countries" />
           { this.geolocationDbNotice() }
-        </React.Fragment>
+        </>
       )
     }
+
+    return null
   }
 
   render() {
@@ -157,4 +161,4 @@ class CountriesMap extends React.Component {
   }
 }
 
-export default withRouter(withThemeConsumer(CountriesMap))
+export default withRouter(withThemeConsumer(Countries))