+++ lib/plausible/stats/base.ex
@@ -117,7 +117,9 @@ defmodule Plausible.Stats.Base do
     "screen" => "screen_size",
     "os" => "operating_system",
     "os_version" => "operating_system_version",
-    "country" => "country_code"
+    "country" => "country_code",
+    "region" => "subdivision1_code",
+    "city" => "city_geoname_id"
   }
 
   def query_sessions(site, query) do