+++ lib/plausible/stats/countries.ex
@@ -1,4 +1,23 @@
 defmodule Plausible.Stats.CountryName do
+  @iso3166_2_source Application.app_dir(:plausible, "priv/iso_3166-2.json")
+  @geonames_source Application.app_dir(:plausible, "priv/cities500.txt")
+
+  @subdivision_names File.read!(@iso3166_2_source)
+                     |> Jason.decode!()
+                     |> Map.get("3166-2")
+                     |> Enum.map(fn %{"code" => code, "name" => name} -> {code, name} end)
+                     |> Enum.into(%{})
+
+  @city_names File.stream!(@geonames_source)
+              |> Stream.map(&String.trim(&1))
+              |> Stream.map(&String.split(&1, "\t"))
+              |> Stream.map(fn [id, name | _rest] -> {String.to_integer(id), name} end)
+              |> Enum.into(%{})
+
+  @city_codes @city_names
+              |> Enum.map(fn {code, name} -> {name, code} end)
+              |> Enum.into(%{})
+
   @country_codes_to_names %{
     "AF" => "Afghanistan",
     "AX" => "Aland Islands",
@@ -505,7 +524,52 @@ defmodule Plausible.Stats.CountryName do
     Map.get(@alpha2_codes, code, code)
   end
 
+  def search_alpha2(name_search_query) do
+    Enum.reduce(@country_codes_to_names, [], fn {code, name}, acc ->
+      matches =
+        name
+        |> String.downcase()
+        |> String.contains?(String.downcase(name_search_query))
+
+      if matches, do: [code | acc], else: acc
+    end)
+  end
+
   def from_iso3166(code) do
     Map.get(@country_codes_to_names, code, code)
+  end
+
+  def from_iso3166_2(code) do
+    Map.get(@subdivision_names, code, code)
+  end
+
+  def search_iso3166_2(name_search_query) do
+    Enum.reduce(@subdivision_names, [], fn {code, name}, acc ->
+      matches =
+        name
+        |> String.downcase()
+        |> String.contains?(String.downcase(name_search_query))
+
+      if matches, do: [code | acc], else: acc
+    end)
+  end
+
+  def search_geoname(name_search_query) do
+    Enum.reduce(@city_names, [], fn {code, name}, acc ->
+      matches =
+        name
+        |> String.downcase()
+        |> String.contains?(String.downcase(name_search_query))
+
+      if matches, do: [code | acc], else: acc
+    end)
+  end
+
+  def from_geoname_id(geoname_id, default) do
+    Map.get(@city_names, geoname_id, default)
+  end
+
+  def to_geoname_id(city_name) do
+    Map.get(@city_codes, city_name)
   end
 end