+++ lib/plausible_web/controllers/api/stats_controller.ex
@@ -516,6 +516,38 @@ defmodule PlausibleWeb.Api.StatsController do
     end
   end
 
+  def regions(conn, params) do
+    site = conn.assigns[:site]
+    query = Query.from(site.timezone, params) |> Filters.add_prefix()
+    pagination = parse_pagination(params)
+
+    countries =
+      Stats.breakdown(site, query, "visit:region", ["visitors"], pagination)
+      |> transform_keys(%{"region" => "code"})
+      |> Enum.map(fn region ->
+        name = Stats.CountryName.from_iso3166_2(region["code"])
+        Map.put(region, "name", name)
+      end)
+
+    json(conn, countries)
+  end
+
+  def cities(conn, params) do
+    site = conn.assigns[:site]
+    query = Query.from(site.timezone, params) |> Filters.add_prefix()
+    pagination = parse_pagination(params)
+
+    cities =
+      Stats.breakdown(site, query, "visit:city", ["visitors"], pagination)
+      |> transform_keys(%{"city" => "code"})
+      |> Enum.map(fn city ->
+        name = Stats.CountryName.from_geoname_id(city["code"], "N/A")
+        Map.put(city, "name", name)
+      end)
+
+    json(conn, cities)
+  end
+
   def browsers(conn, params) do
     site = conn.assigns[:site]
     query = Query.from(site.timezone, params) |> Filters.add_prefix()